move tag related code into its own folder so they are easier to reuse
